+++ src/MailtrapGeneralClient.php
@@ -12,6 +12,7 @@ namespace Mailtrap;
  * @method Api\General\EmailTemplate emailTemplates(int $accountId)
  * @method Api\General\Billing       billing(int $accountId)
  * @method Api\General\ApiToken      apiTokens(int $accountId)
+ * @method Api\General\Organization  organization(int $organizationId)
  *
  * Class MailtrapGeneralClient
  */
@@ -25,5 +26,6 @@ final class MailtrapGeneralClient extends AbstractMailtrapClient
         'emailTemplates' => Api\General\EmailTemplate::class,
         'billing' => Api\General\Billing::class,
         'apiTokens' => Api\General\ApiToken::class,
+        'organization' => Api\General\Organization::class,
     ];
 }