+++ tests/MailtrapGeneralClientTest.php
@@ -29,6 +29,7 @@ class MailtrapGeneralClientTest extends MailtrapClientTestCase
         foreach (MailtrapGeneralClient::API_MAPPING as $key => $item) {
             yield match ($key) {
                 'permissions', 'users', 'contacts', 'emailTemplates', 'billing', 'apiTokens' => [new $item($this->getConfigMock(), self::FAKE_ACCOUNT_ID)],
+                'organization' => [new $item($this->getConfigMock(), self::FAKE_ORGANIZATION_ID)],
                 default => [new $item($this->getConfigMock())],
             };
         }