+++ tests/MailtrapTestCase.php
@@ -16,6 +16,7 @@ abstract class MailtrapTestCase extends TestCase
 {
     public const DEFAULT_API_KEY = 'some_api_key';
     public const FAKE_ACCOUNT_ID = 10001;
+    public const FAKE_ORGANIZATION_ID = 7777;
     public const FAKE_ACCOUNT_ACCESS_ID = 1000001;
     public const FAKE_PROJECT_ID = 2436;
     public const FAKE_INBOX_ID = 4015;